- massives refactoring in den services

+++ Cars/src/de/nak/library/service/impl/AdmonitionProcessServiceImpl.java
@@ -26,33 +26,22 @@ public class AdmonitionProcessServiceImpl implements AdmonitionProcessService {
 		admonitionProcessDAO.save(admonitionProcess);
 	}
 
-//	@Override
-//	public AdmonitionProcess loadAdmonitionProcess(Long admonitionProcessId) {
-//		return admonitionProcessDAO.load(admonitionProcessId);
-//	}
-
 	@Override
 	public void deleteAdmonitionProcess(AdmonitionProcess admonitionProcess) {
-		Set<Admonition> admonitions = admonitionProcess.getAdmonitions();
-		for (Admonition admonition : admonitions)
-			admonitionService.deleteAdmonition(admonition);
-		admonitionProcessDAO.delete(admonitionProcess);
+		if (admonitionProcess != null) {
+			Set<Admonition> admonitions = admonitionProcess.getAdmonitions();
+			for (Admonition admonition : admonitions)
+				admonitionService.deleteAdmonition(admonition);
+			admonitionProcessDAO.delete(admonitionProcess);
+		}
 	}
 
-//	@Override
-//	public List<AdmonitionProcess> loadAllAdmonitionProcesss() {
-//		return admonitionProcessDAO.loadAll();
-//	}
-
-//	@Override
-//	public AdmonitionProcess searchByLending(Lending lending) {
-//		return admonitionProcessDAO.loadByLending(lending);
-//	}
-
-//	@Override
-//	public AdmonitionProcess searchByLendingID(Long lendingId) {
-//		return admonitionProcessDAO.loadByLendingID(lendingId);
-//	}
+	@Override
+	public AdmonitionProcess createAdmonitionProcess() {
+		AdmonitionProcess admonitionProcess = new AdmonitionProcess();
+		admonitionProcess.setAdmonitions(new HashSet<Admonition>());
+		return admonitionProcess;
+	}
 
 	@Override
 	public AdmonitionProcess addAdmonition(AdmonitionProcess admonitionProcess) {
@@ -64,15 +53,6 @@ public class AdmonitionProcessServiceImpl implements AdmonitionProcessService {
 		admonitionProcess.setAdmonitions(admonitions);
 		return admonitionProcess;
 	}
-
-//	@Override
-//	public Integer countAdmonitions(Long admonitionProcessId) {
-//		AdmonitionProcess admonitionProcess = admonitionProcessDAO
-//				.load(admonitionProcessId);
-//		if (admonitionProcess == null)
-//			return null;
-//		return admonitionProcess.getAdmonitions().size();
-//	}
 
 	public void setAdmonitionProcessDAO(
 			AdmonitionProcessDAO admonitionProcessDAO) {
